+++ changelog.d/17279.misc
@@ -1,0 +1,1 @@
+Re-organize Pydantic models and types used in handlers.
