+++ tests/rest/client/test_models.py
@@ -24,7 +24,7 @@ from typing import TYPE_CHECKING
 from typing_extensions import Literal
 
 from synapse._pydantic_compat import HAS_PYDANTIC_V2
-from synapse.rest.client.models import EmailRequestTokenBody
+from synapse.types.rest.client import EmailRequestTokenBody
 
 if TYPE_CHECKING or HAS_PYDANTIC_V2:
     from pydantic.v1 import BaseModel, ValidationError
